+++ packages/web3-core/test/unit/web3_request_manager.test.ts
@@ -688,7 +688,7 @@ describe('Web3RequestManager', () => {
 					err = error;
 				} finally {
 					expect(err).toBeInstanceOf(ResponseError);
-					expect(err.innerError).toEqual(rpcErrorResponse.error);
+					expect(err.cause).toEqual(rpcErrorResponse.error);
 				}
 			});
 		});